pyproject.toml
[tool.poetry]
name = "apibiblioteca"
version = "0.1.0"
description = ""
authors = ["Felifelps <felipefelipe23456@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
quart = "^0.19.3"
firebase-admin = "^6.2.0"
bcrypt = "^4.0.1"
python-dotenv = "^1.0.0"
quart-cors = "^0.7.0"
hypercorn = "^0.14.4"
filelock = "^3.12.4"

[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
